fix(share): ShareButton ref consolidation, dialog close button, barrel export, e2e comment fixes
Suggestions 5–6, 8–12 (PR #664):

Suggestion 5 (ShareButton.tsx): Consolidate scattered useRef declarations into
one group near the top after useState, before useMutation. Import ShareableRegistration
and Tier types for proper ref typing (no eslint-disable, no any).

Suggestion 6 (ShareButton.tsx): Merge two dep-less ref-sync useEffects
(regRef+sharerTierRef and mutateRef) into a single useEffect with explanatory comment.

Suggestion 8 (ShareTriggerDialog.tsx): Add header ✕ close button (aria-label="닫기",
calls onCancel) matching SharePreparingModal pattern. Update ShareTriggerDialog
tests to cover the new button and update focus assertion.

Suggestion 9–10 (e2e/specs/share.spec.ts): Fix SKIPPED/scenario-4 comments —
seed.ts now seeds shared_analyses; expired-row seed is still absent (only that is
missing); scenario 4 is a REAL TEST against E2E_SHARE_ID.

Suggestion 11 (widgets/share/index.ts): Export kindLabel from the share barrel;
update app/share/[id]/page.tsx and opengraph-image.tsx to import from barrel
instead of deep @/widgets/share/lib/kindLabel path.

Suggestion 12 (buildShareSeo.ts): Add alternates: { canonical: null } to both
found and expired/not_found metadata, matching the NOINDEX_SYMBOL_METADATA pattern.
