+++ src/app/share/[id]/opengraph-image.tsx
@@ -1,7 +1,7 @@
 import { OG_IMAGE_HEIGHT, OG_IMAGE_WIDTH } from '@/shared/lib/og';
 import { buildSymbolOgImage } from '@/entities/og-image';
 import { getCachedSharedAnalysis } from '@/entities/shared-analysis/actions/getCachedSharedAnalysis';
-import { kindLabel } from '@/widgets/share/lib/kindLabel';
+import { kindLabel } from '@/widgets/share';
 
 // 공유 스냅샷은 id마다 달라 정적 생성 불가 → force-dynamic
 export const dynamic = 'force-dynamic';
