+++ src/widgets/share/index.ts
@@ -1,1 +1,2 @@
 export { ShareButton } from './ui/ShareButton';
+export { kindLabel } from './lib/kindLabel';
